requirements.txt
sentence-transformers
faiss-cpu
transformers
datasets
torch
numpy
pandas
